requirements.txt
django==2.1.4
honcho==1.0.1
gunicorn==19.9.0
django-htmlmin==0.10.0
django-image-cropping==1.2.0
easy_thumbnails==2.5
WhiteNoise==4.1.2
django-redis==4.10.0
django-ckeditor==5.6.1
Pillow==5.3.0
mysqlclient